+++ lib/segment/src/segment/mod.rs
@@ -3,7 +3,6 @@ mod facet;
 mod formula_rescore;
 pub mod memory;
 mod sampling;
-mod scroll;
 mod search;
 mod segment_ops;
 pub mod vector_data_read;
